chore: move components to middlewares directories
