+++ packages/common/src/mvc/middlewares/AuthenticatedMiddleware.ts
